# contracts/service-uploader/src/generated/models/AssetDto.ts
/* generated using openapi-typescript-codegen -- do not edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
export type AssetDto = {
    /**
     * Virtual folder ID; absent or null means root
     */
    folderId?: string | null;
    _id: string;
    name: string;
    description?: string;
    tags: Array<string>;
    archived: boolean;
    version: number;
    storageStatus: 'AWAITING_UPLOAD' | 'PENDING_STORAGE' | 'READY' | 'STORAGE_FAILED';
    /**
     * Server-generated Spaces object key
     */
    storageKey?: string;
    /**
     * Spaces origin URL; read access depends on object permissions
     */
    storageUrl?: string;
    /**
     * Sanitized and normalized original upload filename
     */
    originalFilename?: string;
    mediaType?: string;
    sizeBytes?: number;
    /**
     * SHA-256 digest of the received file
     */
    checksumSha256?: string;
    receivedAt?: string;
    createdAt: string;
    updatedAt: string;
};