# contracts/service-uploader/src/generated/models/CreateFolderDto.ts
/* generated using openapi-typescript-codegen -- do not edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
export type CreateFolderDto = {
    /**
     * Unique, case-insensitive name
     */
    name: string;
};