# contracts/service-uploader/src/generated/models/FolderDto.ts
/* generated using openapi-typescript-codegen -- do not edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
export type FolderDto = {
    _id: string;
    name: string;
    version: number;
    createdAt: string;
    updatedAt: string;
};