# contracts/service-uploader/src/generated/models/UpdateFolderDto.ts
/* generated using openapi-typescript-codegen -- do not edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
export type UpdateFolderDto = {
    /**
     * Unique, case-insensitive name
     */
    name: string;
};