# openapi.json
{
  "openapi": "3.0.0",
  "paths": {
    "/uploader/folders": {
      "post": {
        "operationId": "FoldersController_create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateFolderDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FolderDto"
                }
              }
            }
          },
          "400": {
            "description": "Malformed ID or invalid folder metadata"
          },
          "401": {
            "description": "Authentication required"
          },
          "409": {
            "description": "Folder name already exists"
          }
        },
        "security": [
          {
            "accessToken": []
          },
          {
            "bearer": []
          }
        ],
        "tags": [
          "folders"
        ]
      },
      "get": {
        "operationId": "FoldersController_findAll",
        "parameters": [],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/FolderDto"
                  }
                }
              }
            }
          },
          "400": {
            "description": "Malformed ID or invalid folder metadata"
          },
          "401": {
            "description": "Authentication required"
          }
        },
        "security": [
          {
            "accessToken": []
          },
          {
            "bearer": []
          }
        ],
        "tags": [
          "folders"
        ]
      }
    },
    "/uploader/folders/{id}": {
      "get": {
        "operationId": "FoldersController_findOne",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FolderDto"
                }
              }
            }
          },
          "400": {
            "description": "Malformed ID or invalid folder metadata"
          },
          "401": {
            "description": "Authentication required"
          },
          "404": {
            "description": "Folder not found"
          }
        },
        "security": [
          {
            "accessToken": []
          },
          {
            "bearer": []
          }
        ],
        "tags": [
          "folders"
        ]
      },
      "patch": {
        "operationId": "FoldersController_update",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateFolderDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FolderDto"
                }
              }
            }
          },
          "400": {
            "description": "Malformed ID or invalid folder metadata"
          },
          "401": {
            "description": "Authentication required"
          },
          "404": {
            "description": "Folder not found"
          },
          "409": {
            "description": "Folder name already exists"
          }
        },
        "security": [
          {
            "accessToken": []
          },
          {
            "bearer": []
          }
        ],
        "tags": [
          "folders"
        ]
      },
      "delete": {
        "operationId": "FoldersController_remove",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": ""
          },
          "400": {
            "description": "Malformed ID or invalid folder metadata"
          },
          "401": {
            "description": "Authentication required"
          },
          "404": {
            "description": "Folder not found"
          },
          "409": {
            "description": "Folder contains assets (including archived)"
          }
        },
        "security": [
          {
            "accessToken": []
          },
          {
            "bearer": []
          }
        ],
        "tags": [
          "folders"
        ]
      }
    },
    "/uploader": {
      "post": {
        "operationId": "UploaderController_create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateAssetDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AssetDto"
                }
              }
            }
          },
          "400": {
            "description": "Invalid asset metadata"
          },
          "401": {
            "description": "Authentication required"
          },
          "404": {
            "description": "Destination folder not found"
          }
        },
        "security": [
          {
            "accessToken": []
          },
          {
            "bearer": []
          }
        ],
        "tags": [
          "uploader"
        ]
      },
      "get": {
        "operationId": "UploaderController_findAll",
        "parameters": [
          {
            "name": "archived",
            "required": false,
            "in": "query",
            "schema": {
              "type": "boolean"
            }
          },
          {
            "name": "folderId",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "root",
            "required": false,
            "in": "query",
            "description": "Only root assets; cannot combine true with folderId",
            "schema": {
              "type": "boolean"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/AssetDto"
                  }
                }
              }
            }
          },
          "400": {
            "description": "Invalid filters or folderId combined with root=true"
          },
          "401": {
            "description": "Authentication required"
          },
          "404": {
            "description": "Filter folder not found"
          }
        },
        "security": [
          {
            "accessToken": []
          },
          {
            "bearer": []
          }
        ],
        "tags": [
          "uploader"
        ]
      }
    },
    "/uploader/upload": {
      "post": {
        "operationId": "UploaderController_upload",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "multipart/form-data": {
              "schema": {
                "type": "object",
                "required": [
                  "file"
                ],
                "properties": {
                  "file": {
                    "type": "string",
                    "format": "binary",
                    "description": "Browser image or video, or PDF (image/*, video/*, or application/pdf)"
                  },
                  "name": {
                    "type": "string",
                    "maxLength": 120
                  },
                  "description": {
                    "type": "string",
                    "maxLength": 2000
                  },
                  "folderId": {
                    "type": "string",
                    "description": "Existing virtual folder ID; omit for root"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "File persisted in Spaces and asset metadata saved as READY",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AssetDto"
                }
              }
            }
          },
          "400": {
            "description": "File is missing, empty, too large, not a browser image/video or PDF, or metadata is invalid"
          },
          "401": {
            "description": "Authentication required"
          },
          "404": {
            "description": "Destination folder not found"
          },
          "409": {
            "description": "Identical content already exists (pending or ready, including archived)"
          },
          "503": {
            "description": "Spaces upload failed; caller must upload the file again"
          }
        },
        "security": [
          {
            "accessToken": []
          },
          {
            "bearer": []
          }
        ],
        "summary": "Persist a file in Spaces and record its asset metadata",
        "tags": [
          "uploader"
        ]
      }
    },
    "/uploader/{id}": {
      "get": {
        "operationId": "UploaderController_findOne",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AssetDto"
                }
              }
            }
          },
          "400": {
            "description": "Malformed asset ID"
          },
          "401": {
            "description": "Authentication required"
          },
          "404": {
            "description": "Asset not found"
          }
        },
        "security": [
          {
            "accessToken": []
          },
          {
            "bearer": []
          }
        ],
        "tags": [
          "uploader"
        ]
      },
      "patch": {
        "operationId": "UploaderController_update",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateAssetDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AssetDto"
                }
              }
            }
          },
          "400": {
            "description": "Malformed ID or invalid metadata"
          },
          "401": {
            "description": "Authentication required"
          },
          "404": {
            "description": "Asset or destination folder not found"
          }
        },
        "security": [
          {
            "accessToken": []
          },
          {
            "bearer": []
          }
        ],
        "tags": [
          "uploader"
        ]
      },
      "delete": {
        "operationId": "UploaderController_remove",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": ""
          },
          "400": {
            "description": "Malformed asset ID"
          },
          "401": {
            "description": "Authentication required"
          },
          "404": {
            "description": "Asset not found"
          }
        },
        "security": [
          {
            "accessToken": []
          },
          {
            "bearer": []
          }
        ],
        "tags": [
          "uploader"
        ]
      }
    },
    "/uploader/{id}/archive": {
      "patch": {
        "operationId": "UploaderController_archive",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AssetDto"
                }
              }
            }
          },
          "400": {
            "description": "Malformed asset ID"
          },
          "401": {
            "description": "Authentication required"
          },
          "404": {
            "description": "Asset not found"
          }
        },
        "security": [
          {
            "accessToken": []
          },
          {
            "bearer": []
          }
        ],
        "tags": [
          "uploader"
        ]
      }
    },
    "/uploader/{id}/unarchive": {
      "patch": {
        "operationId": "UploaderController_unarchive",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AssetDto"
                }
              }
            }
          },
          "400": {
            "description": "Malformed asset ID"
          },
          "401": {
            "description": "Authentication required"
          },
          "404": {
            "description": "Asset not found"
          }
        },
        "security": [
          {
            "accessToken": []
          },
          {
            "bearer": []
          }
        ],
        "tags": [
          "uploader"
        ]
      }
    }
  },
  "info": {
    "title": "NGX service-uploader API",
    "description": "",
    "version": "1.0.0",
    "contact": {}
  },
  "tags": [],
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      },
      "cookie": {
        "type": "apiKey",
        "in": "cookie",
        "name": "accessToken"
      }
    },
    "schemas": {
      "CreateFolderDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "maxLength": 120,
            "description": "Unique, case-insensitive name"
          }
        },
        "required": [
          "name"
        ]
      },
      "FolderDto": {
        "type": "object",
        "properties": {
          "_id": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "version": {
            "type": "number"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time"
          }
        },
        "required": [
          "_id",
          "name",
          "version",
          "createdAt",
          "updatedAt"
        ]
      },
      "UpdateFolderDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "maxLength": 120,
            "description": "Unique, case-insensitive name"
          }
        },
        "required": [
          "name"
        ]
      },
      "CreateAssetDto": {
        "type": "object",
        "properties": {
          "folderId": {
            "type": "string",
            "nullable": true,
            "description": "Virtual folder ID; omit or use null for root"
          },
          "name": {
            "type": "string",
            "maxLength": 120
          },
          "description": {
            "type": "string",
            "maxLength": 2000
          },
          "tags": {
            "maxItems": 50,
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "archived": {
            "type": "boolean"
          }
        },
        "required": [
          "name"
        ]
      },
      "AssetDto": {
        "type": "object",
        "properties": {
          "folderId": {
            "type": "string",
            "nullable": true,
            "description": "Virtual folder ID; absent or null means root"
          },
          "_id": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "description": {
            "type": "string"
          },
          "tags": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "archived": {
            "type": "boolean"
          },
          "version": {
            "type": "number"
          },
          "storageStatus": {
            "type": "string",
            "enum": [
              "AWAITING_UPLOAD",
              "PENDING_STORAGE",
              "READY",
              "STORAGE_FAILED"
            ]
          },
          "storageKey": {
            "type": "string",
            "description": "Server-generated Spaces object key"
          },
          "storageUrl": {
            "type": "string",
            "description": "Spaces origin URL; read access depends on object permissions"
          },
          "originalFilename": {
            "type": "string",
            "description": "Sanitized and normalized original upload filename"
          },
          "mediaType": {
            "type": "string"
          },
          "sizeBytes": {
            "type": "number"
          },
          "checksumSha256": {
            "type": "string",
            "pattern": "^[a-f0-9]{64}$",
            "description": "SHA-256 digest of the received file"
          },
          "receivedAt": {
            "type": "string",
            "format": "date-time"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time"
          }
        },
        "required": [
          "_id",
          "name",
          "tags",
          "archived",
          "version",
          "storageStatus",
          "createdAt",
          "updatedAt"
        ]
      },
      "UpdateAssetDto": {
        "type": "object",
        "properties": {
          "folderId": {
            "type": "string",
            "nullable": true,
            "description": "Virtual folder ID; omit or use null for root"
          },
          "name": {
            "type": "string",
            "maxLength": 120
          },
          "description": {
            "type": "string",
            "maxLength": 2000
          },
          "tags": {
            "maxItems": 50,
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "archived": {
            "type": "boolean"
          }
        }
      }
    }
  }
}